[in progress] refacoring more stuff into NetworkAnchorServer.  The AnchorServer is now more of a Network Anchor Manager, and the AncorClient is now more of a Network Anchor Player.  The renames are coming soon.
